transitionTo and paramsForHandler treat string/number as params
`transitionTo` and `paramsForHandler` now treat string/number
as params that will be passed to model. This is useful for
when you want to call transitionTo, but don't have a model
to provide, don't want to provide one, or want to unify
model loading/resolving with however `model` decides to do
this.