Introduce async support for getHandler
getHandler can now return a Promise for handlers that need to be
fetched asynchronously. Two notable differences from the sync
case are inaccessibleByURL and intermediateTransitionTo.

inaccessibleByURL will now error after going through a loading
state since we need to load the handler before being able to
determine accessibility.

intermediateTransitionTo will assume that the handler is available
synchronously as it should be used for error and loading substates.