Fix queryParamsOnly being correct on transitions.
This property has been set incorrectly on some transitions, notably when the
transition is caused by the Router#refresh() method in queryParamsDidChange.

This is the method ember uses to implement `refreshModel` option of query params,
so fixing this bug allows `refreshModel` to be used in conjunction with
`replace` option for query params, currently it is either-or.